Move tests in package.
