+++ packages/health-computer/tests/tests/mod.rs
@@ -1,0 +1,15 @@
+mod helpers;
+
+mod test_health_scenarios;
+mod test_hls;
+mod test_input_validation;
+mod test_max_borrow_deposit;
+mod test_max_borrow_prop;
+mod test_max_borrow_validation;
+mod test_max_borrow_vault;
+mod test_max_borrow_wallet;
+mod test_max_swap;
+mod test_max_swap_prop;
+mod test_max_swap_validation;
+mod test_max_withdraw;
+mod test_max_withdraw_prop_test;
