+++ packages/health-computer/tests/tests/test_max_borrow_validation.rs
@@ -11,9 +11,7 @@ use mars_types::{
     params::{HlsParams, VaultConfig},
 };
 
-use crate::helpers::{udai_info, umars_info, ustars_info};
-
-pub mod helpers;
+use super::helpers::{udai_info, umars_info, ustars_info};
 
 #[test]
 fn missing_borrow_denom_price_data() {
