+++ packages/health-computer/tests/tests/test_max_swap.rs
@@ -4,20 +4,25 @@ use cosmwasm_std::{coin, Uint128};
 use mars_rover_health_computer::{DenomsData, HealthComputer, VaultsData};
 use mars_types::{
     credit_manager::Positions,
-    health::{AccountKind, BorrowTarget},
+    health::{AccountKind, SwapKind},
 };
 
-use crate::helpers::{udai_info, umars_info};
-
-pub mod helpers;
+use super::helpers::{udai_info, umars_info};
 
 #[test]
-fn max_borrow_wallet_offset_good() {
+fn max_swap_default() {
     let udai = udai_info();
+    let umars = umars_info();
 
     let denoms_data = DenomsData {
-        prices: HashMap::from([(udai.denom.clone(), udai.price)]),
-        params: HashMap::from([(udai.denom.clone(), udai.params.clone())]),
+        prices: HashMap::from([
+            (udai.denom.clone(), udai.price),
+            (umars.denom.clone(), umars.price),
+        ]),
+        params: HashMap::from([
+            (udai.denom.clone(), udai.params.clone()),
+            (umars.denom.clone(), umars.params.clone()),
+        ]),
     };
 
     let vaults_data = VaultsData {
@@ -39,17 +44,24 @@ fn max_borrow_wallet_offset_good() {
     };
 
     let max_borrow_amount =
-        h.max_borrow_amount_estimate(&udai.denom, &BorrowTarget::Wallet).unwrap();
-    assert_eq!(Uint128::new(1014), max_borrow_amount);
+        h.max_swap_amount_estimate(&udai.denom, &umars.denom, &SwapKind::Default).unwrap();
+    assert_eq!(Uint128::new(1200), max_borrow_amount);
 }
 
 #[test]
-fn max_borrow_wallet_offset_margin_of_error() {
+fn max_swap_margin() {
+    let udai = udai_info();
     let umars = umars_info();
 
     let denoms_data = DenomsData {
-        prices: HashMap::from([(umars.denom.clone(), umars.price)]),
-        params: HashMap::from([(umars.denom.clone(), umars.params.clone())]),
+        prices: HashMap::from([
+            (udai.denom.clone(), udai.price),
+            (umars.denom.clone(), umars.price),
+        ]),
+        params: HashMap::from([
+            (udai.denom.clone(), udai.params.clone()),
+            (umars.denom.clone(), umars.params.clone()),
+        ]),
     };
 
     let vaults_data = VaultsData {
@@ -61,7 +73,7 @@ fn max_borrow_wallet_offset_margin_of_error() {
         kind: AccountKind::Default,
         positions: Positions {
             account_id: "123".to_string(),
-            deposits: vec![coin(1200, &umars.denom)],
+            deposits: vec![coin(5000, &udai.denom), coin(500, &umars.denom)],
             debts: vec![],
             lends: vec![],
             vaults: vec![],
@@ -71,8 +83,6 @@ fn max_borrow_wallet_offset_margin_of_error() {
     };
 
     let max_borrow_amount =
-        h.max_borrow_amount_estimate(&umars.denom, &BorrowTarget::Wallet).unwrap();
-
-    // Normally could be 960, but conservative offset rounding has a margin of error
-    assert_eq!(Uint128::new(959), max_borrow_amount);
+        h.max_swap_amount_estimate(&udai.denom, &umars.denom, &SwapKind::Margin).unwrap();
+    assert_eq!(Uint128::new(31351), max_borrow_amount);
 }
