+++ packages/health/tests/tests/mod.rs
@@ -1,0 +1,3 @@
+mod test_from_coins_to_positions;
+mod test_health;
+mod test_health_from_coins;
